Imported hybrid robot
Command-based robot is a periodic (iterated) robot designed to make it
easier for teams to understand and design functional, robust robot code.
It includes a sturdy framework for single-process and multithreaded
(grouped) commands, as well as timed, conditional, and PID command
possibilities.
The hybrid version created here is intended to take all of the benefits
of command-based robot design and apply them to the SampleRobot
framework, eliminating the downsides of iterated program execution and
providing more control over the program.
No testing has been done yet.